+++ server.js
@@ -1,0 +1,3 @@
+// var express = require('express');
+// var app = express();
+// var routes = require('routes/routes.js')(app);
